moved reallocate function to layout.hpp
