+++ domain/include/cstone/domain/layout.hpp
@@ -134,4 +134,28 @@ void relocate(LocalParticleIndex newSize, LocalParticleIndex offset, Arrays&... 
     }
 }
 
+
+//! @brief reallocate arrays to the specified size
+template<class... Arrays>
+void reallocate(std::size_t size, Arrays&... arrays)
+{
+    std::array data{ (&arrays)... };
+
+    size_t current_capacity = data[0]->capacity();
+    if (size > current_capacity)
+    {
+        // limit reallocation growth to 5% instead of 200%
+        auto reserve_size = static_cast<size_t>(double(size) * 1.05);
+        for (auto array : data)
+        {
+            array->reserve(reserve_size);
+        }
+    }
+
+    for (auto array : data)
+    {
+        array->resize(size);
+    }
+}
+
 } // namespace cstone
