+++ domain/test/integration_mpi/exchange_domain.cpp
@@ -33,6 +33,7 @@
 
 #include <gtest/gtest.h>
 
+#include "cstone/domain/layout.hpp"
 #include "cstone/domain/domaindecomp_mpi.hpp"
 
 using namespace cstone;
